objectified-schema/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "objectified-schema-tests"
version = "1.0.0"
description = "SQL schema tests for Objectified"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "pytest>=8.0.0",
    "python-dotenv>=1.0.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

